network cleanup: split uni/windows code
